Bug 2020608 - Part 15: Require unsigned integral types for CeilingLog2 and FindMostSignificantBit. r=sergesanspaille,nbp
Instead of implicitly casting the input to unsigned, require this happens explicitly
in the caller. This aligns with `<bit>` std-functions and `Rotate{Left,Right}` which
also require unsigned inputs.

Differential Revision: https://phabricator.services.mozilla.com/D286324

+++ js/src/jit/RangeAnalysis.cpp
@@ -3591,7 +3591,7 @@ static bool DoesMaskMatchRange(int32_t mask, const Range& range) {
     // Note that the upper bound does not have to be exactly the mask value. For
     // example, consider `x & 0xfff` where `x` is a uint8. That expression can
     // still be optimized to `x`.
-    int bits = 1 + FloorLog2(range.upper());
+    int bits = 1 + FloorLog2(uint32_t(range.upper()));
     uint32_t maskNeeded = (bits == 32) ? 0xffffffff : (uint32_t(1) << bits) - 1;
     if ((mask & maskNeeded) == maskNeeded) {
       return true;
